R codes for analysing Corpus and Experiments
Collocations are understood to be integral building blocks of language processing, alongside individual words, but thus far evidence for the psychological reality of collocations has tended to be confined to English. In contrast to English, Turkish is an agglutinating language, utilizing productive morphology to convey complex meanings using a single word. Given this, we expected Turkish speakers to be less sensitive to phrasal frequencies than English speakers. In Study 1, we conducted a corpus analysis of translation-equivalent adjective–noun collocations (e.g., front door) and found differences between the two languages in frequency counts. In Study 2, we conducted a reaction time experiment to determine the sensitivity of native speakers of English and Turkish to the frequency of adjectives, nouns, and whole collocations. Turkish speakers were less sensitive to whole-phrase frequencies, as predicted, indicating that collocations are processed less holistically in Turkish than English. Both groups demonstrated that processing collocations involves combining information about individual words and phrases. Taken together, we show that speakers are sensitive to frequency information at multiple grain sizes that are attuned to the typology of different languages.